feat(runtime): clean the old runtime integration tests
