+++ pallets/committee/src/lib.rs
@@ -527,9 +527,9 @@ pub mod pallet {
 				let ty = maybe_member.take().ok_or(Error::<T>::NotMember)?;
 
 				// Check if have enough council members
-				if ty == MemberType::Constituent
-					|| Members::<T>::iter_values().filter(|m| *m == MemberType::Council).count()
-						> T::MinCouncilVotes::get()
+				if ty == MemberType::Constituent ||
+					Members::<T>::iter_values().filter(|m| *m == MemberType::Council).count() >
+						T::MinCouncilVotes::get()
 				{
 					VotingEligibility::<T>::take(&member);
 					Ok(ty)
