+++ runtime/integration-tests/src/tests.rs
@@ -1,135 +1,65 @@
 // Copyright 2021 ChainSafe Systems
 // SPDX-License-Identifier: LGPL-3.0-only
 
-use crate::{
-	pint::{self, MockPriceFeed, Runtime as PintRuntime},
-	relay::{self, ProxyType as RelayProxyType, Runtime as RelayRuntime},
-	relay_sovereign_account, sibling_sovereign_account, statemint,
-	types::*,
-	Net, Pint, Relay, Statemint, ADMIN_ACCOUNT, ALICE, INITIAL_BALANCE, PARA_ID, RELAY_CHAIN_ASSET, STATEMINT_PARA_ID,
-};
+use crate::{prelude::*, statemint, util::*};
 use frame_support::{
 	assert_noop, assert_ok,
 	sp_runtime::{traits::Zero, FixedPointNumber},
-	traits::{tokens::fungibles::Inspect, Hooks},
+	traits::tokens::fungibles::Inspect,
 };
+use kusama_runtime::ProxyType as RelayProxyType;
 use orml_traits::MultiCurrency;
-use pallet_price_feed::PriceFeed;
 use pallet_remote_asset_manager::types::StatemintConfig;
-use polkadot_primitives::v1::{AccountId, Balance};
-use primitives::{
-	traits::{MultiAssetRegistry, NavProvider},
-	AssetAvailability,
-};
-use xcm::{
-	v1::{Junction, Junctions, MultiLocation, NetworkId},
-	VersionedMultiAssets, VersionedMultiLocation,
-};
 use xcm_calls::proxy::ProxyType as ParaProxyType;
-use xcm_simulator::TestExt;
-
-type RelayChainPalletXcm = pallet_xcm::Pallet<RelayRuntime>;
-
-#[allow(unused)]
-fn print_events<T: frame_system::Config>(context: &str) {
-	println!("------ {:?} events ------", context);
-	frame_system::Pallet::<T>::events().iter().for_each(|r| {
-		println!("{:?}", r.event);
-	});
-}
-
-#[allow(unused)]
-fn run_to_block<Runtime>(n: u32)
-where
-	Runtime: pallet_remote_asset_manager::Config<BlockNumber = BlockNumber>,
-{
-	while frame_system::Pallet::<Runtime>::block_number() < n {
-		pallet_remote_asset_manager::Pallet::<Runtime>::on_finalize(frame_system::Pallet::<Runtime>::block_number());
-		frame_system::Pallet::<Runtime>::on_finalize(frame_system::Pallet::<Runtime>::block_number());
-		frame_system::Pallet::<Runtime>::set_block_number(frame_system::Pallet::<Runtime>::block_number() + 1);
-		frame_system::Pallet::<Runtime>::on_initialize(frame_system::Pallet::<Runtime>::block_number());
-		pallet_remote_asset_manager::Pallet::<Runtime>::on_initialize(frame_system::Pallet::<Runtime>::block_number());
-	}
-}
-
-/// registers the relay chain as liquid asset
-fn register_relay() {
-	// prepare index fund so NAV is available
-	let deposit = 1_000;
-	assert_ok!(orml_tokens::Pallet::<PintRuntime>::deposit(RELAY_CHAIN_ASSET, &ADMIN_ACCOUNT, 1_000));
-	assert_ok!(pallet_asset_index::Pallet::<PintRuntime>::register_asset(
-		pint::Origin::signed(ADMIN_ACCOUNT),
-		RELAY_CHAIN_ASSET,
-		AssetAvailability::Liquid(MultiLocation::parent()),
-	));
-	assert_ok!(pallet_asset_index::Pallet::<PintRuntime>::add_asset(
-		pint::Origin::signed(ADMIN_ACCOUNT),
-		RELAY_CHAIN_ASSET,
-		deposit,
-		deposit
-	));
-	assert!(pallet_asset_index::Pallet::<PintRuntime>::is_liquid_asset(&RELAY_CHAIN_ASSET));
-}
-
-/// transfer the given amount of relay chain currency into the account on the
-/// parachain
-fn transfer_to_para(relay_deposit_amount: Balance, who: AccountId) {
-	Relay::execute_with(|| {
-		// transfer from relay to parachain
-		assert_ok!(RelayChainPalletXcm::reserve_transfer_assets(
-			relay::Origin::signed(who.clone()),
-			Box::new(VersionedMultiLocation::V1(Junctions::X1(Junction::Parachain(PARA_ID)).into())),
-			Box::new(VersionedMultiLocation::V1(
-				Junctions::X1(Junction::AccountId32 { network: NetworkId::Any, id: who.clone().into() }).into()
-			)),
-			Box::new(VersionedMultiAssets::V1((Junctions::Here, relay_deposit_amount).into())),
-			0,
-		));
-	});
-	Pint::execute_with(|| {
-		// ensure deposit arrived
-		assert_eq!(orml_tokens::Pallet::<PintRuntime>::balance(RELAY_CHAIN_ASSET, &who), relay_deposit_amount);
-	});
-}
+use xcm_emulator::TestExt;
 
 #[test]
 fn para_account_funded_on_relay() {
 	Net::reset();
 
-	Relay::execute_with(|| {
-		let para_balance_on_relay = pallet_balances::Pallet::<RelayRuntime>::free_balance(&relay_sovereign_account());
+	Kusama::execute_with(|| {
+		let para_balance_on_relay = pallet_balances::Pallet::<KusamaRuntime>::free_balance(&relay_sovereign_account());
 		assert_eq!(para_balance_on_relay, INITIAL_BALANCE);
 	});
 }
 
 #[test]
 fn can_deposit_from_relay() {
+	use pallet_price_feed::PriceFeed;
+
 	Net::reset();
-	Pint::execute_with(|| register_relay());
-	let deposit = 1_000;
+	Shot::execute_with(|| register_relay());
+	let deposit = 1_000_000_000;
 	transfer_to_para(deposit, ALICE);
 
-	Pint::execute_with(|| {
-		let initial_index_tokens = pallet_asset_index::Pallet::<PintRuntime>::index_token_issuance();
-		let index_token_balance = pallet_asset_index::Pallet::<PintRuntime>::index_token_balance(&ALICE);
-		let nav = pallet_asset_index::Pallet::<PintRuntime>::nav().unwrap();
+	Shot::execute_with(|| {
+		let initial_index_tokens = pallet_asset_index::Pallet::<ShotRuntime>::index_token_issuance();
+		let index_token_balance = pallet_asset_index::Pallet::<ShotRuntime>::index_token_balance(&ALICE);
+
+		// create feed
+		create_and_submit_feed(ADMIN_ACCOUNT, RELAY_CHAIN_ASSET, 1);
+
+		let nav = pallet_asset_index::Pallet::<ShotRuntime>::nav().unwrap();
 
 		// alice has 1000 units of relay chain currency in her account on the parachain
-		assert_ok!(pallet_asset_index::Pallet::<PintRuntime>::deposit(
-			pint::Origin::signed(ALICE),
+		assert_ok!(pallet_asset_index::Pallet::<ShotRuntime>::deposit(
+			committee_origin(ALICE).into(),
 			RELAY_CHAIN_ASSET,
 			deposit
 		));
 		// no more relay chain assets
-		assert!(orml_tokens::Pallet::<PintRuntime>::balance(RELAY_CHAIN_ASSET, &ALICE).is_zero());
+		assert!(orml_tokens::Pallet::<ShotRuntime>::balance(RELAY_CHAIN_ASSET, &ALICE).is_zero());
 
-		let deposit_value = MockPriceFeed::get_price(RELAY_CHAIN_ASSET).unwrap().checked_mul_int(deposit).unwrap();
+		let deposit_value = pallet_price_feed::Pallet::<ShotRuntime>::get_price(RELAY_CHAIN_ASSET)
+			.unwrap()
+			.checked_mul_int(deposit)
+			.unwrap();
 		let received = nav.reciprocal().unwrap().saturating_mul_int(deposit_value);
 		assert_eq!(
-			pallet_asset_index::Pallet::<PintRuntime>::index_token_balance(&ALICE),
+			pallet_asset_index::Pallet::<ShotRuntime>::index_token_balance(&ALICE),
 			received + index_token_balance
 		);
-		assert_eq!(pallet_asset_index::Pallet::<PintRuntime>::index_token_issuance(), received + initial_index_tokens);
+		assert_eq!(pallet_asset_index::Pallet::<ShotRuntime>::index_token_issuance(), received + initial_index_tokens);
 	});
 }
 
@@ -137,57 +67,59 @@ fn can_deposit_from_relay() {
 fn can_transact_register_proxy() {
 	Net::reset();
 
-	Pint::execute_with(|| {
+	Shot::execute_with(|| {
 		register_relay();
-		assert_ok!(pallet_remote_asset_manager::Pallet::<PintRuntime>::send_add_proxy(
-			pint::Origin::signed(ADMIN_ACCOUNT),
+		assert_ok!(pallet_remote_asset_manager::Pallet::<ShotRuntime>::send_add_proxy(
+			shot_runtime::Origin::signed(ADMIN_ACCOUNT),
 			RELAY_CHAIN_ASSET,
 			ParaProxyType(RelayProxyType::Staking as u8),
 			Option::None
 		));
 
 		assert_noop!(
-			pallet_remote_asset_manager::Pallet::<PintRuntime>::send_add_proxy(
-				pint::Origin::signed(ADMIN_ACCOUNT),
+			pallet_remote_asset_manager::Pallet::<ShotRuntime>::send_add_proxy(
+				shot_runtime::Origin::signed(ADMIN_ACCOUNT),
 				RELAY_CHAIN_ASSET,
 				ParaProxyType(RelayProxyType::Staking as u8),
 				Option::None
 			),
-			pallet_remote_asset_manager::Error::<PintRuntime>::AlreadyProxy
+			pallet_remote_asset_manager::Error::<ShotRuntime>::AlreadyProxy
 		);
 	});
 
-	Relay::execute_with(|| {
+	Kusama::execute_with(|| {
 		// verify the proxy is registered
 		let proxy =
-			pallet_proxy::Pallet::<RelayRuntime>::find_proxy(&relay_sovereign_account(), &ADMIN_ACCOUNT, Option::None)
+			pallet_proxy::Pallet::<KusamaRuntime>::find_proxy(&relay_sovereign_account(), &ADMIN_ACCOUNT, Option::None)
 				.unwrap();
 		assert_eq!(proxy.proxy_type, RelayProxyType::Staking);
 	});
 }
 
 #[test]
-fn can_transact_staking() {
+fn tcan_transact_staking() {
+	env_logger::init();
 	Net::reset();
 	// `- 1` for avoiding dust account issue
 	//
 	// see also https://github.com/open-web3-stack/open-runtime-module-library/issues/427
-	let bond = 10_000 - 1;
+	let bond = 1_000_000_000 - 1;
+	let deposit = 2_000_000_000;
 
-	Pint::execute_with(|| {
+	Shot::execute_with(|| {
 		register_relay();
 		// mint some funds first to cover the transfer
-		assert_ok!(pint::Currency::deposit(RELAY_CHAIN_ASSET, &ADMIN_ACCOUNT, 1_000_000));
+		assert_ok!(shot_runtime::Currencies::deposit(RELAY_CHAIN_ASSET, &ADMIN_ACCOUNT, deposit));
 
 		// fails to bond extra, no initial bond
 		assert_noop!(
-			pallet_remote_asset_manager::Pallet::<PintRuntime>::do_send_bond_extra(RELAY_CHAIN_ASSET, bond,),
-			pallet_remote_asset_manager::Error::<PintRuntime>::NotBonded
+			pallet_remote_asset_manager::Pallet::<ShotRuntime>::do_send_bond_extra(RELAY_CHAIN_ASSET, bond),
+			pallet_remote_asset_manager::Error::<ShotRuntime>::NotBonded
 		);
 
 		// transact a bond call that adds `ADMIN_ACCOUNT` as controller
-		assert_ok!(pallet_remote_asset_manager::Pallet::<PintRuntime>::send_bond(
-			pint::Origin::signed(ADMIN_ACCOUNT),
+		assert_ok!(pallet_remote_asset_manager::Pallet::<ShotRuntime>::send_bond(
+			shot_runtime::Origin::signed(ADMIN_ACCOUNT),
 			RELAY_CHAIN_ASSET,
 			ADMIN_ACCOUNT.into(),
 			bond,
@@ -195,30 +127,30 @@ fn can_transact_staking() {
 		));
 
 		assert_noop!(
-			pallet_remote_asset_manager::Pallet::<PintRuntime>::send_bond(
-				pint::Origin::signed(ADMIN_ACCOUNT),
+			pallet_remote_asset_manager::Pallet::<ShotRuntime>::send_bond(
+				shot_runtime::Origin::signed(ADMIN_ACCOUNT),
 				RELAY_CHAIN_ASSET,
 				ADMIN_ACCOUNT.into(),
 				bond,
 				xcm_calls::staking::RewardDestination::Staked
 			),
-			pallet_remote_asset_manager::Error::<PintRuntime>::AlreadyBonded
+			pallet_remote_asset_manager::Error::<ShotRuntime>::AlreadyBonded
 		);
 	});
 
-	Relay::execute_with(|| {
+	Kusama::execute_with(|| {
 		// make sure `ADMIN_ACCOUNT` is now registered as controller
-		let ledger = pallet_staking::Ledger::<RelayRuntime>::get(&ADMIN_ACCOUNT).unwrap();
+		let ledger = pallet_staking::Ledger::<KusamaRuntime>::get(&ADMIN_ACCOUNT).unwrap();
 		assert_eq!(ledger.total, bond);
 	});
 
-	Pint::execute_with(|| {
+	Shot::execute_with(|| {
 		// bond extra
-		assert_ok!(pallet_remote_asset_manager::Pallet::<PintRuntime>::do_send_bond_extra(RELAY_CHAIN_ASSET, bond));
+		assert_ok!(pallet_remote_asset_manager::Pallet::<ShotRuntime>::do_send_bond_extra(RELAY_CHAIN_ASSET, bond));
 	});
 
-	Relay::execute_with(|| {
-		let ledger = pallet_staking::Ledger::<RelayRuntime>::get(&ADMIN_ACCOUNT).unwrap();
+	Kusama::execute_with(|| {
+		let ledger = pallet_staking::Ledger::<KusamaRuntime>::get(&ADMIN_ACCOUNT).unwrap();
 		// bond + 1x bond_extra
 		assert_eq!(ledger.total, 2 * bond);
 	});
@@ -248,53 +180,53 @@ fn can_transfer_to_statemint() {
 	});
 
 	let transfer_amount = 1_000;
-	Pint::execute_with(|| {
+	Shot::execute_with(|| {
 		// try to send PINT, but no config yet
 		assert_noop!(
-			pallet_remote_asset_manager::Pallet::<PintRuntime>::transfer_to_statemint(
-				pint::Origin::signed(ALICE),
+			pallet_remote_asset_manager::Pallet::<ShotRuntime>::transfer_to_statemint(
+				shot_runtime::Origin::signed(ALICE),
 				transfer_amount
 			),
-			pallet_remote_asset_manager::Error::<PintRuntime>::NoStatemintConfigFound
+			pallet_remote_asset_manager::Error::<ShotRuntime>::NoStatemintConfigFound
 		);
 
 		let config = StatemintConfig { parachain_id: STATEMINT_PARA_ID, enabled: false };
 
-		assert_ok!(pallet_remote_asset_manager::Pallet::<PintRuntime>::set_statemint_config(
-			pint::Origin::signed(ADMIN_ACCOUNT),
+		assert_ok!(pallet_remote_asset_manager::Pallet::<ShotRuntime>::set_statemint_config(
+			shot_runtime::Origin::signed(ADMIN_ACCOUNT),
 			config
 		));
 
 		// not enabled yet
 		assert_noop!(
-			pallet_remote_asset_manager::Pallet::<PintRuntime>::transfer_to_statemint(
-				pint::Origin::signed(ALICE),
+			pallet_remote_asset_manager::Pallet::<ShotRuntime>::transfer_to_statemint(
+				shot_runtime::Origin::signed(ALICE),
 				transfer_amount
 			),
-			pallet_remote_asset_manager::Error::<PintRuntime>::StatemintDisabled
+			pallet_remote_asset_manager::Error::<ShotRuntime>::StatemintDisabled
 		);
 
-		assert_ok!(pallet_remote_asset_manager::Pallet::<PintRuntime>::enable_statemint_xcm(pint::Origin::signed(
-			ADMIN_ACCOUNT
-		)));
+		assert_ok!(pallet_remote_asset_manager::Pallet::<ShotRuntime>::enable_statemint_xcm(
+			shot_runtime::Origin::signed(ADMIN_ACCOUNT)
+		));
 
 		// // no funds to transfer from empty account
 		// assert_noop!(
-		// 	pallet_remote_asset_manager::Pallet::<PintRuntime>::transfer_to_statemint(
-		// 		pint::Origin::signed(EMPTY_ACCOUNT),
+		// 	pallet_remote_asset_manager::Pallet::<ShotRuntime>::transfer_to_statemint(
+		// 		shot_runtime::Origin::signed(EMPTY_ACCOUNT),
 		// 		transfer_amount
 		// 	),
-		// 	pallet_balances::Error::<PintRuntime>::InsufficientBalance
+		// 	pallet_balances::Error::<ShotRuntime>::InsufficientBalance
 		// );
 		//
-		// pallet_remote_asset_manager::Pallet::<PintRuntime>::transfer_to_statemint(
-		// 	pint::Origin::signed(ALICE),
+		// pallet_remote_asset_manager::Pallet::<ShotRuntime>::transfer_to_statemint(
+		// 	shot_runtime::Origin::signed(ALICE),
 		// 	transfer_amount
 		// );
 
 		// transfer from pint -> statemint to mint SPINT
-		// assert_ok!(pallet_remote_asset_manager::Pallet::<PintRuntime>::transfer_to_statemint(
-		// 	pint::Origin::signed(ALICE),
+		// assert_ok!(pallet_remote_asset_manager::Pallet::<ShotRuntime>::transfer_to_statemint(
+		// 	shot_runtime::Origin::signed(ALICE),
 		// 	transfer_amount
 		// ));
 	});
